# docs/swagger.yaml
basePath: /
definitions:
  github_com_solumD_WBTech_L0_internal_model.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    type: object
  github_com_solumD_WBTech_L0_internal_model.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    type: object
  github_com_solumD_WBTech_L0_internal_model.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/github_com_solumD_WBTech_L0_internal_model.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_solumD_WBTech_L0_internal_model.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/github_com_solumD_WBTech_L0_internal_model.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    type: object
  github_com_solumD_WBTech_L0_internal_model.Payment:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: integer
      delivery_cost:
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    type: object
  handler.getOrderByUIDResponse:
    properties:
      error:
        type: string
      order:
        $ref: '#/definitions/github_com_solumD_WBTech_L0_internal_model.Order'
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: WBTech_L0 order service
  version: "1.0"
paths:
  /order/{uid}:
    get:
      description: gets order by its' uid
      operationId: get-order
      parameters:
      - description: order's uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getOrderByUIDResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: GetOrder
      tags:
      - order
swagger: "2.0"